webapi/requirements.txt
Flask==3.0.2
Flask_Cors==4.0.0
Flask_MySQLdb==2.0.0
Flask_SocketIO==5.3.6
openai==1.14.2
pymongo==4.6.2
python-dotenv==1.0.1
python_docx==1.1.0
tencentcloud_sdk_python==3.0.1116
tencentcloud_sdk_python_common==3.0.1137
tencentcloud_sdk_python_sms==3.0.1137
gunicorn
gevent